[feat]일기보기 버튼으로 수정, tailwind로 변경
